requirements.txt
# Python dependencies for BGRemover & Image Upscaler
# Make sure you're using Python 3.10

# Core image processing libraries
rembg>=2.0.0
realesrgan>=0.3.0

# Additional dependencies that may be needed
Pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.0
basicsr>=1.4.2

# Development dependencies (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0 

gfpgan 

torch
torchvision
Pillow
gdown 